Init node only once

+++ src/main/java/org/vaadin/artur/exampledata/NodeScriptInterface.java
@@ -2,7 +2,6 @@ package org.vaadin.artur.exampledata;
 
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -14,8 +13,13 @@ public class NodeScriptInterface {
 
     private static final DateTimeFormatter OUTPUT_DATE_PATTERN = DateTimeFormatter.ofPattern("M/d/y");
     private static NodeUtil node = new NodeUtil();
+    private static boolean inited = false;
 
     private static void init() throws InterruptedException, IOException {
+        if (inited) {
+            return;
+        }
+        inited = true;
         File chanceTempFile = File.createTempFile("exampleutil-bundle", "js");
         chanceTempFile.deleteOnExit();
         try (FileOutputStream out = new FileOutputStream(chanceTempFile)) {
